Geant4DD4hep: split DD4hep tools into separate module
